+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 18.7.0 Fixes
 
 - `[Datagrid]` Fixed a bug in validate cell when using lookup. ([#4838](https://inforwiki.atlassian.net/browse/IDS-4838))
+- `[Datepicker]` Updated example page with french localization. ([#4608](https://inforwiki.atlassian.net/browse/IDS-4608))
 
 ## 18.6.0